+++ module/test/action/PastellCoreTestExternalData.class.php
@@ -4,7 +4,7 @@ class PastellCoreTestExternalData extends ChoiceActionExecutor{
 
 	public function go(){
 
-		$recuperateur = new Recuperateur($_POST);
+		$recuperateur = $this->getRecuperateur();
 		$choix = $recuperateur->get('choix');
 
 		$donneesFormulaire = $this->getDonneesFormulaire();